Перенос кода в отдельные классы.
