+++ tests/ConsoleTest/CriticalSectionTests.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace ConsoleTest
+{
+    class CriticalSectionTests
+    {
+    }
+}
